PTEID changes for 2048 bit keys and SHA256 hash
PtEid card can sign with 1024 or 2048 bit RSA
as long as the data to be signed is less the 36 bytes.

V3 cards can handle SHA256 hash and digest if it strips off the
digest header before sending the stripped hash to the card, along
with a flag to have the card add the SHA256 digest header before padding.
Thus it can handle SHA256 and stay within the 36 byte limit.

gensafe_flags2algref will combine the env->algorithm_flags and
env->algorithm_flags_original to see if a hash with digest header is being used.

If no hash is is indicated, it will look at env->in_data and env->in_data
by using sc_pkcs1_strip_digest_info_prefix to see if the calling application
may be providing a SHA256 hash and digest created externally for OpenSC.

This commit depends on  "Improved handling of Hashes"

 Date:      Mon Feb 27 07:45:25 2017 -0600
 On branch ImproveHashHandling-2
 Changes to be committed:
	modified:   src/libopensc/card-gemsafeV1.c